+++ lib/rubocop/cop/style/avoid_global_vars.rb
@@ -5,6 +5,10 @@ module Rubocop
     module Style
       # This cops looks for uses of global variables.
       # It does not report offences for built-in global variables.
+      # Built-in global variables are allowed by default. Additionally
+      # users can allow additional variables via the AllowedVariables option.
+      #
+      # Note that backreferences like $1, $2, etc are not global variables.
       class AvoidGlobalVars < Cop
         MSG = 'Do not introduce global variables.'
 
@@ -36,7 +40,20 @@ module Rubocop
           $stdin $stdout $stderr
           $DEBUG $FILENAME $VERBOSE $SAFE
           $-0 $-a $-d $-F $-i $-I $-l $-p $-v $-w
+          $CLASSPATH $JRUBY_VERSION $JRUBY_REVISION $ENV_JAVA
         ).map(&:to_sym)
+
+        def user_vars
+          if cop_config['AllowedVariables']
+            cop_config['AllowedVariables'].map(&:to_sym)
+          else
+            {}
+          end
+        end
+
+        def allowed_var?(global_var)
+          BUILT_IN_VARS.include?(global_var) || user_vars.include?(global_var)
+        end
 
         def on_gvar(node)
           check(node)
@@ -49,7 +66,7 @@ module Rubocop
         def check(node)
           global_var, = *node
 
-          convention(node, :name) unless BUILT_IN_VARS.include?(global_var)
+          convention(node, :name) unless allowed_var?(global_var)
         end
       end
     end